refactor: move template record helpers into TemplateProcessor
